+++ gtest/src/lib.rs
@@ -2,6 +2,7 @@ mod log;
 mod manager;
 mod program;
 mod system;
+mod mailbox;
 
 pub use log::{CoreLog, Log, RunResult};
 pub use program::{calculate_program_id, Program, WasmProgram};
